Add chapter folder name number prefixes
