+++ src/applications/pensions/config/chapters/02-military-history/generalHistory.js
