+++ src/applications/pensions/config/chapters/02-military-history/servicePeriods.js
