+++ src/applications/pensions/config/chapters/03-health-and-employment-information/vaTreatmentHistory.js
