+++ src/applications/pensions/config/form.js
@@ -55,28 +55,28 @@ import monthlyIncomeUI from '../definitions/monthlyIncome';
 import expectedIncomeUI from '../definitions/expectedIncome';
 import { additionalSourcesSchema } from '../definitions/additionalSources';
 import otherExpensesUI from '../definitions/otherExpenses';
-import applicantInformation from './chapters/applicant-information/applicantInformation';
-import mailingAddress from './chapters/applicant-information/mailingAddress';
-import contactInformation from './chapters/applicant-information/contactInformation';
-import servicePeriods from './chapters/military-history/servicePeriods';
-import generalHistory from './chapters/military-history/generalHistory';
-import pow from './chapters/military-history/pow';
-import age from './chapters/health-and-employment-information/age';
-import socialSecurityDisability from './chapters/health-and-employment-information/socialSecurityDisability';
-import medicaidCoverage from './chapters/health-and-employment-information/medicaidCoverage';
-import medicaidStatus from './chapters/health-and-employment-information/medicaidStatus';
-import medicalCondition from './chapters/health-and-employment-information/medicalCondition';
-import nursingHome from './chapters/health-and-employment-information/nursingHome';
-import specialMonthlyPension from './chapters/health-and-employment-information/specialMonthlyPension';
-import vaTreatmentHistory from './chapters/health-and-employment-information/vaTreatmentHistory';
-import federalTreatmentHistory from './chapters/health-and-employment-information/federalTreatmentHistory';
-import generateMedicalCentersSchemas from './chapters/health-and-employment-information/medicalCenters';
-import currentEmployment from './chapters/health-and-employment-information/currentEmployment';
-import generateEmployersSchemas from './chapters/health-and-employment-information/employmentHistory';
-import maritalStatus from './chapters/household-information/maritalStatus';
-import currentSpouse from './chapters/household-information/currentSpouse';
-import dateOfCurrentMarriage from './chapters/household-information/dateOfCurrentMarriage';
-import reasonForCurrentSeparation from './chapters/household-information/reasonForCurrentSeparation';
+import applicantInformation from './chapters/01-applicant-information/applicantInformation';
+import mailingAddress from './chapters/01-applicant-information/mailingAddress';
+import contactInformation from './chapters/01-applicant-information/contactInformation';
+import servicePeriods from './chapters/02-military-history/servicePeriods';
+import generalHistory from './chapters/02-military-history/generalHistory';
+import pow from './chapters/02-military-history/pow';
+import age from './chapters/03-health-and-employment-information/age';
+import socialSecurityDisability from './chapters/03-health-and-employment-information/socialSecurityDisability';
+import medicaidCoverage from './chapters/03-health-and-employment-information/medicaidCoverage';
+import medicaidStatus from './chapters/03-health-and-employment-information/medicaidStatus';
+import medicalCondition from './chapters/03-health-and-employment-information/medicalCondition';
+import nursingHome from './chapters/03-health-and-employment-information/nursingHome';
+import specialMonthlyPension from './chapters/03-health-and-employment-information/specialMonthlyPension';
+import vaTreatmentHistory from './chapters/03-health-and-employment-information/vaTreatmentHistory';
+import federalTreatmentHistory from './chapters/03-health-and-employment-information/federalTreatmentHistory';
+import generateMedicalCentersSchemas from './chapters/03-health-and-employment-information/medicalCenters';
+import currentEmployment from './chapters/03-health-and-employment-information/currentEmployment';
+import generateEmployersSchemas from './chapters/03-health-and-employment-information/employmentHistory';
+import maritalStatus from './chapters/04-household-information/maritalStatus';
+import currentSpouse from './chapters/04-household-information/currentSpouse';
+import dateOfCurrentMarriage from './chapters/04-household-information/dateOfCurrentMarriage';
+import reasonForCurrentSeparation from './chapters/04-household-information/reasonForCurrentSeparation';
 
 import { validateAfterMarriageDate } from '../validation';
 import migrations from '../migrations';
