+++ src/applications/pensions/tests/chapters/01-applicant-information/applicantInformation.unit.spec.jsx
@@ -18,7 +18,7 @@ import formConfig from '../../../config/form';
 import applicantInformation, {
   isOver65,
   setDefaultIsOver65,
-} from '../../../config/chapters/applicant-information/applicantInformation';
+} from '../../../config/chapters/01-applicant-information/applicantInformation';
 
 const definitions = formConfig.defaultDefinitions;
