+++ src/applications/pensions/tests/chapters/03-health-and-employment-information/treatmentHistory.unit.spec.jsx
@@ -3,8 +3,8 @@ import {
   testNumberOfWebComponentFields,
 } from '../pageTests.spec';
 import formConfig from '../../../config/form';
-import treatmentHistory from '../../../config/chapters/health-and-employment-information/vaTreatmentHistory';
-import generateMedicalCentersSchemas from '../../../config/chapters/health-and-employment-information/medicalCenters';
+import treatmentHistory from '../../../config/chapters/03-health-and-employment-information/vaTreatmentHistory';
+import generateMedicalCentersSchemas from '../../../config/chapters/03-health-and-employment-information/medicalCenters';
 
 describe('pension treatment history page', () => {
   it('should render web components in form correctly', () => {
